refactor(api): extract upload routes
Refs #331

Refs #358

Original route extraction from #358 by @shadowinlife.

Maintainer integration narrows the first API modularization step to the upload and Shadow report routes so no unused duplicate API modules land in main.
